pyproject.toml
[tool.poetry]
name = "rottenpotatoes"
version = "0.1.0"
description = ""
authors = ["Gabriel Rodrigues <gabrielsr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.3"
bootstrap-flask = "^2.2.0"
flask-wtf = "^1.1.1"
flask-sqlalchemy = "^3.0.3"
flask-login = "^0.6.2"
flask-migrate = "^4.0.4"
flask-bcrypt = "^1.0.1"
email-validator = "^2.0.0.post2"
python-dateutil = "^2.8.2"


[tool.poetry.group.test]  # This part can be left out

[tool.poetry.group.test.dependencies]
pytest = "^6.0.0"
pytest-mock = "*"


[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
mkdocs = "*"
mkdocs-material = "^9.1.6"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pylint = "^2.17.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


